refactor: move FieldPathNavigator to field_path_detection module
- Move FieldPathNavigator from lazy_config.py to field_path_detection.py
- Update all imports to use the new location
- Improve code organization by grouping field path utilities together

FieldPathNavigator is a utility for navigating dot-separated field paths
and logically belongs with FieldPathDetector in the field_path_detection
module rather than in lazy_config.py.

All functionality remains unchanged - this is a pure refactoring.
