+++ openhcs/core/lazy_config.py
@@ -85,46 +85,6 @@ def _create_static_field_extractor(obj):
 # Removed single-use _create_resolution_config method - inlined at call site
 
 
-class FieldPathNavigator:
-    """Utility for navigating dot-separated field paths in object hierarchies."""
-
-    @staticmethod
-    def navigate_to_instance(current_global_config: Any, field_path: Optional[str] = None) -> Optional[Any]:
-        """
-        Navigate to instance using explicit field path.
-
-        Args:
-            current_global_config: Thread-local storage object or global config instance
-            field_path: Dot-separated path to navigate (None = root)
-
-        Returns:
-            Instance at the specified field path, or None if not found
-        """
-        # Handle both thread-local storage objects and direct config instances
-        if hasattr(current_global_config, "value"):
-            if not current_global_config.value:
-                return None
-            instance = current_global_config.value
-        else:
-            # Direct config instance
-            instance = current_global_config
-
-        if field_path is None:
-            # Root instance - return the global config directly
-            return instance
-
-        # Navigate dot-separated path
-        for field in field_path.split('.'):
-            if instance is None:
-                return None
-            # Use object.__getattribute__ to avoid triggering lazy resolution during navigation
-            try:
-                instance = object.__getattribute__(instance, field)
-            except AttributeError:
-                return None
-
-        return instance
-
 
 @dataclass(frozen=True)
 class ResolutionConfig:
@@ -222,7 +182,7 @@ class LazyMethodBindings:
 
                                         for field_path in field_paths:
                                             try:
-                                                from openhcs.core.lazy_config import FieldPathNavigator
+                                                from openhcs.core.field_path_detection import FieldPathNavigator
                                                 context_value = FieldPathNavigator.navigate_to_instance(current_context, field_path)
                                                 if context_value and hasattr(context_value, name):
                                                     global_value = getattr(context_value, name)
